# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST API documentation",
    "description": "A voting system for deciding where to have lunch. Course TopJava</a> (Graduation project solution)",
    "contact": {
      "name": "Andrey Kozlov",
      "email": "avgur33@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
...
    "/api/admin/restaurants/{id}": {
      "get": {
        "tags": [
          "AdminRestaurantController"
        ],
        "summary": "Get restaurant by restaurant id",
        "operationId": "get_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of restaurant that needs to be fetched. Use 1 for testing.",
            "required": true,
            "content": {
              "*/*": {
                "example": 1
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The restaurant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurant"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          },
          "423": {
            "description": "Locked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminRestaurantController"
        ],
        "summary": "Update restaurant",
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of restaurant that needs to be deleted. Use 1 for testing.",
            "required": true,
            "content": {
              "*/*": {
                "example": 1
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Restaurant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Update the restaurant"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdminRestaurantController"
        ],
        "summary": "Delete restaurant",
        "operationId": "delete_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of restaurant that needs to be deleted. Use 1 for testing.",
            "required": true,
            "content": {
              "*/*": {
                "example": 1
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/restaurants": {
      "get": {
        "tags": [
          "AdminRestaurantController"
        ],
        "summary": "Get all restaurants",
        "operationId": "getAll_1",
        "responses": {
          "200": {
            "description": "List of restaurants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Restaurant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "AdminRestaurantController"
        ],
        "summary": "Create restaurant",
        "operationId": "creatWithLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Restaurant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Create the restaurant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurant"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/restaurants/{restaurantId}/menu": {
      "get": {
        "tags": [
          "AdminMenuController"
        ],
        "summary": "Get All menus for the restaurant",
        "description": "You can use filter between Start date and End Date inclusive",
        "operationId": "getAll_2",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "description": "The id of restaurant. Use 1 for testing.",
            "required": true,
            "content": {
              "*/*": {
                "example": 1
              }
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date. Format yyyy-MM-dd.",
            "required": false,
            "content": {
              "*/*": {
                "example": "2020-02-21"
              }
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date. Format yyyy-MM-dd.",
            "required": false,
            "content": {
              "*/*": {
                "example": "2022-02-21"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Menu"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminMenuController"
        ],
        "summary": "Create menu for the restaurant",
        "description": "Number of dishes must be between 2 and 5",
        "operationId": "creatWithLocation_1",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "description": "The id of restaurant. Use 3 for testing.",
            "required": true,
            "content": {
              "*/*": {
                "example": 3
              }
            }
          },
          {
            "name": "forDate",
            "in": "query",
            "description": "For date. Format yyyy-MM-dd.",
            "required": false,
            "content": {
              "*/*": {
                "example": "2022-02-21"
              }
            }
          },
          {
            "name": "dishes",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Menu"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/restaurants/{restaurantId}/dishes": {
      "get": {
        "tags": [
          "AdminDishController"
        ],
        "summary": "Get dishes for restaurant",
        "operationId": "getAll_3",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "description": "The id of restaurant. Use 1 for testing.",
            "required": true,
            "content": {
              "*/*": {
                "example": 1
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The dishes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dish"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminDishController"
        ],
        "summary": "Create dish for restaurant",
        "operationId": "creatWithLocation_2",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "description": "The id of restaurant. Use 1 for testing.",
            "required": true,
            "content": {
              "*/*": {
                "example": 1
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dish"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created dish for the restaurant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dish"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          },
          "423": {
            "description": "Locked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorInfo": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Dish": {
        "required": [
          "name",
          "price"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID for new entity must be null",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string",
            "description": "name",
            "example": "newName"
          },
          "price": {
            "type": "integer",
            "description": "min = 10",
            "format": "int32",
            "example": 100
          }
        }
      },
      "Restaurant": {
        "required": [
          "location",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID for new entity must be null",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 2,
            "type": "string",
            "description": "name",
            "example": "newName"
          },
          "location": {
            "maxLength": 100,
            "minLength": 5,
            "type": "string",
            "description": "name",
            "example": "Moscow City"
          }
        }
      },
...
      "Menu": {
        "required": [
          "forDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID for new entity must be null",
            "format": "int32"
          },
          "forDate": {
            "type": "string",
            "description": "For this date",
            "format": "date",
            "example": "2020-02-20"
          },
          "dishes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dish"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}